* commit '5a4c622b565b273cf23fd18f7aff96975fafb7a9': (22 commits)
  [EDELIVER-4866] changes as per pull request
  [EDELIVER-4866] build form scratch 4.1.1 multitenancy changhelog file
  [EDELIVER-4866] start 4.1.1 changelog from scratch
  [EDELIVER-4866] revert back
  [EDELIVER-4866] revert back
  [EDELIVER-4866] revert back
  [EDELIVER-4866] move changes to delta model
  [EDELIVER-4866] revert back
  [EDELIVER-4866] revert back
  [EDELIVER-4866] revert back
  [EDELIVER-4866] revert back
  [EDELIVER-4866] revert back
  [EDELIVER-4866] revert back
  [EDELIVER-4866] fixed some skipped columns
  [EDELIVER-4866] revert back 4.0 multitenancy changes
  [EDELIVER-4866] added multi tenancy updates
  [EDELIVER-4866] fixed all MySQL errors
  [EDELIVER-4866] fixed TB_EVENT_ALERT
  [EDELIVER-4866] fix TB_EVENT_PROPERTY table
  [EDELIVER-4866] added migration script for all tables from 4.0 version
  ...

+++ Domibus-MSH/src/main/java/eu/domibus/core/alerts/dao/AlertDao.java
@@ -6,7 +6,6 @@ import eu.domibus.core.alerts.model.common.AlertCriteria;
 import eu.domibus.core.alerts.model.persist.*;
 import eu.domibus.logging.DomibusLoggerFactory;
 import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Repository;
 
 import javax.persistence.Query;
@@ -43,7 +42,7 @@ public class AlertDao extends BasicDao<Alert> {
         //create root entity specifying that we want to eager fetch. (Avoid the N+1 hibernate problem)
         final Root<Alert> root = criteria.from(Alert.class);
 
-        final Subquery<Integer> subQuery = criteria.subquery(Integer.class);
+        final Subquery<Long> subQuery = criteria.subquery(Long.class);
         Root<Alert> subRoot = subQuery.from(Alert.class);
 
         //Do first a subQuery to retrieve the filtered alerts id based on criteria.
@@ -134,7 +133,7 @@ public class AlertDao extends BasicDao<Alert> {
         //create root entity specifying that we want to eager fetch. (Avoid the N+1 hibernate problem)
         final Root<Alert> root = criteria.from(Alert.class);
 
-        final Subquery<Integer> subQuery = criteria.subquery(Integer.class);
+        final Subquery<Long> subQuery = criteria.subquery(Long.class);
         Root<Alert> subRoot = subQuery.from(Alert.class);
 
         //Do first a subQuery to retrieve the filtered alerts id based on criteria.
